+++ app/src/main/java/com/practicum/playlistmaker/player/data/repository/PlayerRepositoryImpl.kt
@@ -1,6 +1,5 @@
 package com.practicum.playlistmaker.player.data.repository
 
-
 import android.media.MediaPlayer
 import android.util.Log
 import com.practicum.playlistmaker.player.domain.repository.PlayerRepository
@@ -15,7 +14,10 @@ class PlayerRepositoryImpl @Inject constructor() : PlayerRepository {
     private var isPrepared: Boolean = false
     private var completionListener: (() -> Unit)? = null
 
-
+    /**
+     * Инициализирует MediaPlayer и подготавливает аудиофайл к воспроизведению.
+     * Устанавливает обработчики ошибок, готовности и завершения воспроизведения.
+     */
     override suspend fun prepare(url: String?) {
         try {
             mediaPlayer = MediaPlayer().apply {
@@ -29,7 +31,6 @@ class PlayerRepositoryImpl @Inject constructor() : PlayerRepository {
 
                 setOnPreparedListener {
                     isPrepared = true
-                    // После подготовки устанавливаем сохранённую позицию
                     if (savedPosition > 0L) {
                         seekTo(savedPosition.toInt())
                     }
@@ -50,14 +51,20 @@ class PlayerRepositoryImpl @Inject constructor() : PlayerRepository {
         }
     }
 
+    /**
+     * Запускает воспроизведение с сохранённой позиции (если есть) или с начала.
+     */
     override suspend fun play() {
         playWithPosition(null)
     }
 
+    /**
+     * Запускает воспроизведение с указанной позиции или сохранённой позиции.
+     * Если медиаплеер ещё не готов, ожидает готовности перед воспроизведением.
+     */
     override suspend fun playWithPosition(resumePosition: Long?) {
         mediaPlayer?.let { player ->
             if (!isPrepared) {
-                // Если не подготовлен — ждём подготовки
                 player.setOnPreparedListener {
                     resumeWithPosition(player, resumePosition)
                 }
@@ -67,6 +74,10 @@ class PlayerRepositoryImpl @Inject constructor() : PlayerRepository {
         } ?: Log.e("PlayerRepository", "MediaPlayer is null in play()")
     }
 
+    /**
+     * Вспомогательный метод для запуска воспроизведения с заданной позиции.
+     * Выполняет seekTo() при необходимости и запускает воспроизведение.
+     */
     private fun resumeWithPosition(player: MediaPlayer, resumePosition: Long?) {
         val targetPosition = resumePosition ?: savedPosition
         if (targetPosition > 0L) {
@@ -77,19 +88,27 @@ class PlayerRepositoryImpl @Inject constructor() : PlayerRepository {
         Log.d("PlayerRepository", "play() called, isPlaying=${player.isPlaying}, position=${player.currentPosition}")
     }
 
+    /**
+     * Приостанавливает воспроизведение и сохраняет текущую позицию.
+     */
     override suspend fun pause() {
         mediaPlayer?.pause()
-        // Сохраняем текущую позицию при паузе
         mediaPlayer?.currentPosition?.toLong()?.let { savedPosition = it }
     }
 
+    /**
+     * Останавливает воспроизведение и сохраняет текущую позицию.
+     * По сути, работает аналогично pause(), но  более полную остановку.
+     */
     override suspend fun stop() {
         mediaPlayer?.pause()
-        // Не сбрасываем позицию на 0 — сохраняем для возобновления
         mediaPlayer?.currentPosition?.toLong()?.let { savedPosition = it }
-        // Не вызываем reset() — оставляем подготовленным
     }
 
+    /**
+     * Полностью сбрасывает состояние медиаплеера: освобождает ресурсы,
+     * обнуляет сохранённую позицию и флаги состояния.
+     */
     override suspend fun reset() {
         mediaPlayer?.release()
         mediaPlayer = null
@@ -98,19 +117,36 @@ class PlayerRepositoryImpl @Inject constructor() : PlayerRepository {
         isPrepared = false
     }
 
+    /**
+     * Проверяет, воспроизводится ли аудио в данный момент.
+     * @return true, если медиаплеер воспроизводит аудио, иначе false.
+     */
     override fun isPlaying(): Boolean {
         return mediaPlayer?.isPlaying ?: false
     }
 
+    /**
+     * Возвращает текущую позицию воспроизведения в миллисекундах.
+     * Если медиаплеер недоступен, возвращает сохранённую позицию.
+     * @return текущая позиция воспроизведения или сохранённая позиция.
+     */
     override fun getCurrentPosition(): Long {
         return mediaPlayer?.currentPosition?.toLong() ?: savedPosition
     }
 
+    /**
+     * Устанавливает слушатель завершения воспроизведения.
+     * Сохраняет лямбду для вызова при завершении и привязывает её к MediaPlayer.
+     */
     override fun setOnCompletionListener(listener: () -> Unit) {
         completionListener = listener
         mediaPlayer?.setOnCompletionListener { listener() }
     }
 
+    /**
+     * Перемещает воспроизведение на указанную позицию в миллисекундах.
+     * Обновляет сохранённую позицию после seekTo().
+     */
     override fun seekTo(position: Long) {
         mediaPlayer?.seekTo(position.toInt())
         savedPosition = position
