+++ app/src/main/java/com/practicum/playlistmaker/player/domain/DelayedTrackActionUseCase.kt
@@ -5,10 +5,20 @@ import com.practicum.playlistmaker.core.contract.DelayedTrackActionUseCaseContra
 import com.practicum.playlistmaker.core.models.Track
 import javax.inject.Inject
 
+/**
+ * UseCase для выполнения действия с треком после заданной задержки.
+ * Используется, например, для добавления трека в историю после короткого ожидания.
+ */
 class DelayedTrackActionUseCase @Inject constructor(
     private val delayProvider: DelayProvider
 ) : DelayedTrackActionUseCaseContract {
 
+    /**
+     * Выполняет действие с треком после указанной задержки.
+     * @param track трек, с которым нужно выполнить действие
+     * @param delayMillis время задержки в миллисекундах
+     * @param onDelayedAction действие, которое будет выполнено после задержки
+     */
     override suspend operator fun invoke(
         track: Track,
         delayMillis: Long,
