+++ app/src/main/java/com/practicum/playlistmaker/player/domain/PreparePlaybackUseCase.kt
@@ -4,10 +4,20 @@ import com.practicum.playlistmaker.player.domain.repository.PlayerRepository
 import com.practicum.playlistmaker.core.contract.PreparePlaybackUseCaseContract
 import javax.inject.Inject
 
+/**
+ * UseCase для подготовки аудиофайла к воспроизведению.
+ * Инициализирует медиаплеер и подготавливает трек по указанному URL.
+ */
 class PreparePlaybackUseCase @Inject constructor(
     private val playerRepository: PlayerRepository
 ) : PreparePlaybackUseCaseContract {
 
+    /**
+     * Подготавливает аудиофайл к воспроизведению по указанному URL.
+     * @param previewUrl URL аудиофайла для воспроизведения
+     * @return Result.success(Unit) при успешной подготовке,
+     *         Result.failure(e) при ошибке
+     */
     override suspend operator fun invoke(previewUrl: String?): Result<Unit> {
         return try {
             playerRepository.prepare(previewUrl)
