+++ app/src/main/java/com/practicum/playlistmaker/player/domain/SetPlaybackCompletionListenerUseCase.kt
@@ -4,10 +4,18 @@ import com.practicum.playlistmaker.player.domain.repository.PlayerRepository
 import com.practicum.playlistmaker.core.contract.SetPlaybackCompletionListenerUseCaseContract
 import javax.inject.Inject
 
+/**
+ * UseCase для установки слушателя события завершения воспроизведения.
+ * Позволяет реагировать на окончание воспроизведения аудио.
+ */
 class SetPlaybackCompletionListenerUseCase @Inject constructor(
     private val playerRepository: PlayerRepository
 ) : SetPlaybackCompletionListenerUseCaseContract {
 
+    /**
+     * Устанавливает слушатель события завершения воспроизведения.
+     * @param onCompletion функция, которая будет вызвана при завершении воспроизведения
+     */
     override suspend fun invoke(onCompletion: () -> Unit) {
         playerRepository.setOnCompletionListener(onCompletion)
     }
