+++ app/src/main/java/com/practicum/playlistmaker/player/domain/StopPlaybackUseCase.kt
@@ -4,10 +4,19 @@ import com.practicum.playlistmaker.player.domain.repository.PlayerRepository
 import com.practicum.playlistmaker.core.contract.StopPlaybackUseCaseContract
 import javax.inject.Inject
 
+/**
+ * UseCase для остановки воспроизведения аудио.
+ * Приостанавливает медиаплеер и сохраняет текущую позицию.
+ */
 class StopPlaybackUseCase @Inject constructor(
     private val playerRepository: PlayerRepository
 ) : StopPlaybackUseCaseContract {
 
+    /**
+     * Останавливает воспроизведение аудио.
+     * @return Result.success(Unit) при успешной остановке,
+     *         Result.failure(e) при ошибке
+     */
     override suspend fun invoke(): Result<Unit> {
         return try {
             playerRepository.stop()
@@ -17,4 +26,3 @@ class StopPlaybackUseCase @Inject constructor(
         }
     }
 }
-
