+++ app/src/main/java/com/practicum/playlistmaker/player/domain/TogglePlaybackUseCase.kt
@@ -4,10 +4,23 @@ import com.practicum.playlistmaker.player.domain.repository.PlayerRepository
 import com.practicum.playlistmaker.core.contract.TogglePlaybackUseCaseContract
 import javax.inject.Inject
 
+/**
+ * UseCase для переключения состояния воспроизведения: play ↔ pause.
+ * Если воспроизведение не идёт — запускает его.
+ * Если идёт — приостанавливает.
+ */
 class TogglePlaybackUseCase @Inject constructor(
     private val playerRepository: PlayerRepository
 ) : TogglePlaybackUseCaseContract {
 
+    /**
+     * Переключает состояние воспроизведения:
+     * - если не играло: запускает воспроизведение ;
+     * - если играло: приостанавливает воспроизведение.
+     * @param seekPosition позиция для seekTo перед запуском (опционально);
+     * @return Result.success(isPlaying) — текущее состояние воспроизведения после операции;
+     *         Result.failure(e) — ошибка при выполнении.
+     */
     override suspend operator fun invoke(seekPosition: Long?): Result<Boolean> {
         return try {
             val wasPlaying = playerRepository.isPlaying()
