+++ app/src/main/java/com/practicum/playlistmaker/player/domain/repository/PlayerRepository.kt
@@ -6,7 +6,7 @@ interface PlayerRepository {
     suspend fun pause()
     suspend fun stop()
     suspend fun reset()
-    suspend fun playWithPosition(resumePosition: Long?) // новый метод с позицией
+    suspend fun playWithPosition(resumePosition: Long?)
     fun seekTo(position: Long)
     fun isPlaying(): Boolean
     fun getCurrentPosition(): Long
