+++ app/src/main/java/com/practicum/playlistmaker/player/ui/AlbumViewHolder.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.search.ui.viewholder
+package com.practicum.playlistmaker.player.ui
 
 import android.util.Log
 import android.view.View
@@ -10,11 +10,22 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.practicum.playlistmaker.R
-import com.practicum.playlistmaker.core.models.Track
 import com.practicum.playlistmaker.core.contract.FormatTrackDurationUseCaseContract
+import com.practicum.playlistmaker.core.models.Track
 import com.practicum.playlistmaker.core.utils.DateFormatter
 
-
+/**
+ * ViewHolder для отображения трека в режиме детального просмотра (аудиоплеер).
+ * Отображает полную информацию о треке: обложку, название, исполнителя, длительность,
+ * дату релиза, жанр, страну, а также кнопки управления воспроизведением и действиями.
+ *
+ * @param itemView корневой View элемента RecyclerView
+ * @param onClickListener обработчик клика по всему элементу трека
+ * @param onPlayButtonClick обработчик нажатия кнопки воспроизведения/паузы
+ * @param onAddToPlaylistClick обработчик добавления трека в плейлист
+ * @param onFavoriteClick обработчик отметки трека как избранного
+ * @param formatDurationUseCase use case для форматирования длительности трека
+ */
 class AlbumViewHolder(
     itemView: View,
     private val onClickListener: (Track) -> Unit,
@@ -24,6 +35,7 @@ class AlbumViewHolder(
     private val formatDurationUseCase: FormatTrackDurationUseCaseContract
 ) : RecyclerView.ViewHolder(itemView) {
 
+    // UI‑компоненты элемента
     private val albumImageView: ImageView = itemView.findViewById(R.id.album)
     private val textTrackName: TextView = itemView.findViewById(R.id.textTrackName)
     private val textArtistName: TextView = itemView.findViewById(R.id.textArtistName)
@@ -37,9 +49,21 @@ class AlbumViewHolder(
     private val plusButton: Button = itemView.findViewById(R.id.ic_button_plus)
     private val likeButton: Button = itemView.findViewById(R.id.ic_button_like)
 
+    /** Утилита для форматирования даты релиза */
     private val dateFormatter = DateFormatter()
+
+    /** Текущий трек, привязанный к этому ViewHolder */
     private var currentTrack: Track? = null
 
+    /**
+     * Привязывает данные трека к UI‑элементам и обновляет состояние кнопки воспроизведения.
+     * Загружает обложку через Glide с закруглёнными углами.
+     *
+     * @param track данные трека для отображения
+     * @param isPlaying флаг, указывающий, воспроизводится ли трек сейчас
+     * @param currentTimeMillis текущая позиция воспроизведения в миллисекундах
+     * @param formattedTime форматированное время воспроизведения (например, «03:45»)
+     */
     fun bind(
         track: Track,
         isPlaying: Boolean,
@@ -48,6 +72,7 @@ class AlbumViewHolder(
     ) {
         currentTrack = track
 
+        // Заполняем текстовые поля данными трека
         textTrackName.text = track.trackName
         textArtistName.text = track.artistName
         releaseDateTextView.text = dateFormatter.formatReleaseDate(track.releaseDate)
@@ -55,11 +80,13 @@ class AlbumViewHolder(
         primaryGenreNameTextView.text = track.primaryGenreName
         countryTextView.text = track.country
 
+        // Отображаем текущее время воспроизведения и общую длительность трека
         timeTextView.text = formattedTime
         trackTimeMillisTextView.text = track.trackTimeMillis?.let {
             formatDurationUseCase.invoke(it)
         } ?: ""
 
+        // Загружаем обложку альбома через Glide
         val cornerRadiusPx = (8 * itemView.resources.displayMetrics.density).toInt()
         if (track.artworkUrl100 != null) {
             Glide.with(itemView.context)
@@ -77,8 +104,12 @@ class AlbumViewHolder(
         setupClickListeners(track)
     }
 
-
-
+    /**
+     * Настраивает обработчики кликов для всех интерактивных элементов ViewHolder.
+     * Обработчики делегируют события наружу через лямбды, переданные в конструктор.
+     *
+     * @param track текущий трек для передачи в обработчики
+     */
     private fun setupClickListeners(track: Track) {
         itemView.setOnClickListener { onClickListener(track) }
         playButton.setOnClickListener { onPlayButtonClick(track) }
@@ -86,6 +117,13 @@ class AlbumViewHolder(
         likeButton.setOnClickListener { onFavoriteClick(track) }
     }
 
+    /**
+     * Обновляет иконку кнопки воспроизведения в зависимости от состояния воспроизведения.
+     * Показывает иконку паузы, если трек играет, или иконку воспроизведения, если на паузе.
+     * Логирует изменение состояния для отладки.
+     *
+     * @param isPlaying флаг воспроизведения (играет/пауза)
+     */
     fun updatePlayButtonState(isPlaying: Boolean) {
         Log.d("AlbumViewHolder", "updatePlayButtonState: isPlaying=$isPlaying")
         playButton.setImageResource(
