+++ app/src/main/java/com/practicum/playlistmaker/search/data/repository/ItunesRepositoryImpl.kt
@@ -7,11 +7,28 @@ import com.practicum.playlistmaker.search.data.mapper.SearchResponseMapper
 import com.practicum.playlistmaker.search.data.network.ItunesApi
 import javax.inject.Inject
 
+/**
+ * Реализация репозитория для взаимодействия с iTunes API.
+ * Выполняет поиск треков и преобразует ответ API в доменную модель.
+ *
+ * @param api клиент для запросов к iTunes API
+ * @param searchResponseMapper маппер для преобразования DTO в доменную модель
+ */
 class ItunesRepositoryImpl @Inject constructor(
     private val api: ItunesApi,
     private val searchResponseMapper: SearchResponseMapper
 ) : ItunesRepository {
 
+    /**
+     * Выполняет поиск треков по запросу через iTunes API.
+     * - отправляет запрос к API;
+     * - преобразует полученный DTO-объект в доменную модель SearchResponse;
+     * - возвращает результат для дальнейшего использования в бизнес‑логике.
+     *
+     * @param query поисковый запрос (название трека, исполнителя и т. д.)
+     * @return объект SearchResponse с результатами поиска
+     * @throws Exception если запрос к API завершился ошибкой
+     */
     override suspend fun search(query: String): SearchResponse {
         val response: SearchResponseDTO = api.searchTracks(query)
         return searchResponseMapper.toDomain(response)
