+++ app/src/main/java/com/practicum/playlistmaker/search/domain/GetSearchHistoryUseCase.kt
@@ -5,10 +5,23 @@ import com.practicum.playlistmaker.search.domain.repository.HistoryRepository
 import com.practicum.playlistmaker.core.contract.GetSearchHistoryUseCaseContract
 import javax.inject.Inject
 
-class GetSearchHistoryUseCase @Inject constructor (
+/**
+ * UseCase для получения истории поиска.
+ * Извлекает сохранённые треки из локального хранилища.
+ *
+ * @param historyRepository репозиторий для работы с историей поиска
+ */
+class GetSearchHistoryUseCase @Inject constructor(
     private val historyRepository: HistoryRepository
 ) : GetSearchHistoryUseCaseContract {
 
+    /**
+     * Получает историю поиска из локального хранилища.
+     * Возвращает список треков в порядке добавления (последний — первый).
+     *
+     * @return список треков из истории поиска;
+     *         пустой список, если история пуста или произошла ошибка
+     */
     override suspend operator fun invoke(): List<Track> {
         return historyRepository.getHistory()
     }
