+++ app/src/main/java/com/practicum/playlistmaker/search/ui/SearchActivity.kt
@@ -15,7 +15,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.practicum.playlistmaker.R
 import com.practicum.playlistmaker.core.models.Track
 import com.practicum.playlistmaker.player.ui.AudioPlayerActivity
-import com.practicum.playlistmaker.search.ui.adapter.TrackAdapter
+import com.practicum.playlistmaker.search.ui.adapter.SearchTrackAdapter
 import com.practicum.playlistmaker.search.ui.parcel.toParcelable
 import com.practicum.playlistmaker.core.constants.Constants.Companion.SEARCH_QUERY_KEY
 import com.practicum.playlistmaker.core.constants.Constants.Companion.VIEW_TYPE_TRACK
@@ -23,14 +23,24 @@ import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-import javax.inject.Inject
 
+/**
+ * Активность для поиска треков и отображения истории поиска.
+ * Реализует функционал:
+ * - поиск треков с задержкой ввода (debounce 2 с);
+ * - отображение результатов поиска, ошибок и состояния загрузки;
+ * - управление историей поиска (просмотр, очистка);
+ * - навигация к аудиоплееру при клике на трек.
+ *
+ * Использует ViewModel для управления состоянием и взаимодействия с бизнес‑логикой.
+ */
 @AndroidEntryPoint
 class SearchActivity : AppCompatActivity() {
 
+    /** ViewModel для управления поиском и историей */
     private val viewModel: SearchViewModel by viewModels()
 
-    // Вьюшки
+    // UI‑компоненты активности
     private lateinit var backTextView: TextView
     private lateinit var searchEditText: EditText
     private lateinit var resetButton: ImageView
@@ -43,11 +53,11 @@ class SearchActivity : AppCompatActivity() {
     private lateinit var historyRecyclerViewKit: LinearLayout
     private lateinit var progressBar: ProgressBar
 
-    // Адаптеры
-    private lateinit var tracksAdapter: TrackAdapter
-    private lateinit var historyAdapter: TrackAdapter
+    // Адаптеры для RecyclerView
+    private lateinit var tracksAdapter: SearchTrackAdapter
+    private lateinit var historyAdapter: SearchTrackAdapter
 
-    // Данные
+    // Вспомогательные данные
     private var searchQuery: String = ""
     private var clickJob: Job? = null
 
@@ -63,6 +73,12 @@ class SearchActivity : AppCompatActivity() {
         viewModel.loadHistory()
     }
 
+    /**
+     * Инициализирует и настраивает все UI‑компоненты:
+     * - находит вью по ID;
+     * - создаёт и настраивает адаптеры для списков треков и истории;
+     * - устанавливает LayoutManager для RecyclerView.
+     */
     private fun initViews() {
         backTextView = findViewById(R.id.back)
         searchEditText = findViewById(R.id.search_edit_text)
@@ -76,33 +92,33 @@ class SearchActivity : AppCompatActivity() {
         historyRecyclerViewKit = findViewById(R.id.search_history_layout)
         progressBar = findViewById(R.id.progressBar)
 
-        tracksAdapter = TrackAdapter(
+        // Адаптер для результатов поиска
+        tracksAdapter = SearchTrackAdapter(
             tracks = emptyList(),
-            viewType = VIEW_TYPE_TRACK,
-            onTrackClick = { track ->
-                viewModel.onTrackClicked(track)
-            },
-            onClickPlayButton = {},
-            onAddToPlaylist = {},
-            onFavorite = {},
+            onTrackClick = { track -> viewModel.onTrackClicked(track) },
             formatDurationUseCase = viewModel.formatTrackDurationUseCase
         )
+
         recyclerView.adapter = tracksAdapter
         recyclerView.layoutManager = LinearLayoutManager(this)
 
-        historyAdapter = TrackAdapter(
+        // Адаптер для истории поиска
+        historyAdapter = SearchTrackAdapter(
             tracks = emptyList(),
-            viewType = VIEW_TYPE_TRACK,
             onTrackClick = { track -> openAudioPlayer(track) },
-            onClickPlayButton = {},
-            onAddToPlaylist = {},
-            onFavorite = {},
             formatDurationUseCase = viewModel.formatTrackDurationUseCase
         )
         historyRecyclerView.adapter = historyAdapter
         historyRecyclerView.layoutManager = LinearLayoutManager(this)
     }
 
+    /**
+     * Настраивает обработчики кликов для кнопок:
+     * - «назад» — закрывает активность;
+     * - сброс — очищает поле поиска, скрывает клавиатуру, обновляет UI;
+     * - обновление — повторяет поиск при ошибке;
+     * - очистка истории — вызывает очистку истории в ViewModel.
+     */
     private fun setupClickListeners() {
         backTextView.setOnClickListener { finish() }
         resetButton.setOnClickListener {
@@ -117,14 +133,19 @@ class SearchActivity : AppCompatActivity() {
             }
         }
 
-
         clearHistoryButton.setOnClickListener {
             viewModel.clearHistory()
-            historyRecyclerViewKit.visibility = View.GONE // Восстанавливаем видимость после очистки
+            historyRecyclerViewKit.visibility = View.GONE // Скрываем блок истории после очистки
         }
-
     }
 
+    /**
+     * Настраивает наблюдение за вводом текста в поле поиска:
+     * - отображает/скрывает кнопку сброса;
+     * - обновляет видимость истории;
+     * - реализует debounce (2 с) перед отправкой запроса;
+     * - фильтрует локальные треки при пустом запросе.
+     */
     private fun setupTextWatchers() {
         var searchJob: Job? = null
         searchEditText.doOnTextChanged { text, _, _, _ ->
@@ -150,6 +171,12 @@ class SearchActivity : AppCompatActivity() {
         }
     }
 
+    /**
+     * Восстанавливает состояние активности из сохранённых данных (например, при повороте экрана).
+     * Устанавливает текст поиска и запускает поиск, если запрос не пуст.
+     *
+     * @param savedInstanceState пакет с сохранённым состоянием
+     */
     private fun restoreState(savedInstanceState: Bundle?) {
         if (savedInstanceState != null) {
             searchQuery = savedInstanceState.getString(SEARCH_QUERY_KEY, "")
@@ -158,6 +185,12 @@ class SearchActivity : AppCompatActivity() {
         }
     }
 
+    /**
+     * Подписывается на изменения состояния ViewModel:
+     * - searchState: отображает загрузку, результаты, ошибки, отсутствие результатов;
+     * - historyState: обновляет список истории и её видимость;
+     * - trackToOpen: открывает аудиоплеер для выбранного трека.
+     */
     private fun observeViewModel() {
         viewModel.searchState.observe(this) { state ->
             when (state) {
@@ -182,7 +215,7 @@ class SearchActivity : AppCompatActivity() {
             when (state) {
                 HistoryState.Loading -> {}
                 HistoryState.Empty -> {
-                    historyAdapter.updateList(emptyList())
+                    historyAdapter            .updateList(emptyList())
                     updateHistoryVisibility()
                 }
                 is HistoryState.HistoryLoaded -> {
@@ -190,7 +223,6 @@ class SearchActivity : AppCompatActivity() {
                     updateHistoryVisibility()
                 }
                 HistoryState.HistoryCleared -> {
-                    // Исправление: было View.VISIBLE, теперь View.GONE
                     historyRecyclerViewKit.visibility = View.GONE
                     updateHistoryVisibility()
                 }
@@ -206,15 +238,23 @@ class SearchActivity : AppCompatActivity() {
         }
     }
 
-
+    /**
+     * Обновляет список треков в адаптере и управляет видимостью RecyclerView и макета «нет результатов».
+     *
+     * @param tracks список треков для отображения
+     */
     private fun updateTracksList(tracks: List<Track>) {
         tracksAdapter.updateList(tracks)
         recyclerView.visibility = if (tracks.isNotEmpty()) View.VISIBLE else View.GONE
         showNoResults(tracks.isEmpty() && searchQuery.isNotEmpty())
     }
 
-
-
+    /**
+     * Открывает экран аудиоплеера для выбранного трека.
+     * Передаёт трек через Intent с использованием Parcelable.
+     *
+     * @param track трек для воспроизведения
+     */
     private fun openAudioPlayer(track: Track) {
         val intent = Intent(this, AudioPlayerActivity::class.java)
         val parcelableTrack = track.toParcelable()
@@ -222,6 +262,9 @@ class SearchActivity : AppCompatActivity() {
         startActivity(intent)
     }
 
+    /**
+     * Отображает индикатор загрузки и скрывает другие элементы интерфейса.
+     */
     private fun showLoading() {
         progressBar.visibility = View.VISIBLE
         recyclerView.visibility = View.INVISIBLE
@@ -229,32 +272,49 @@ class SearchActivity : AppCompatActivity() {
         errorLayout.visibility = View.INVISIBLE
     }
 
-
+    /**
+     * Скрывает индикатор загрузки.
+     */
     private fun hideLoading() {
         progressBar.visibility = View.INVISIBLE  // Просто скрываем прогресс‑бар
     }
 
+    /**
+     * Отображает макет с сообщением об ошибке сети/API.
+     */
     private fun showError() {
         errorLayout.visibility = View.VISIBLE
         noResultsLayout.visibility = View.GONE
     }
 
-
+    /**
+     * Управляет видимостью макета «нет результатов» в зависимости от условий.
+     * Показывает, если:
+     * - нет результатов;
+     * - запрос не пуст;
+     * - последний поиск не завершился ошибкой.
+     *
+     * @param show флаг, указывающий, нужно ли показать макет
+     */
     private fun showNoResults(show: Boolean) {
         noResultsLayout.visibility = if (show && !viewModel.isLastSearchFailed) View.VISIBLE else View.GONE
         errorLayout.visibility = View.GONE
     }
 
-
-
-
-
-
+    /**
+     * Скрывает виртуальную клавиатуру.
+     */
     private fun hideKeyboard() {
         val imm = getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
         imm.hideSoftInputFromWindow(searchEditText.windowToken, 0)
     }
 
+    /**
+     * Обновляет видимость блока истории поиска в зависимости от:
+     * - наличия фокуса на поле поиска;
+     * - пустоты запроса;
+     * - наличия записей в истории.
+     */
     private fun updateHistoryVisibility() {
         val isEmptyQuery = searchEditText.text.isEmpty()
         val hasFocus = searchEditText.hasFocus()
