+++ app/src/main/java/com/practicum/playlistmaker/settings/domain/GetThemeStateUseCase.kt
@@ -1,13 +1,25 @@
+
 package com.practicum.playlistmaker.settings.domain
 
 import com.practicum.playlistmaker.settings.domain.repository.SettingsRepository
 import com.practicum.playlistmaker.core.contract.GetThemeStateUseCaseContract
 import javax.inject.Inject
 
+/**
+ * Use case для получения текущего состояния темы приложения (тёмная/светлая).
+ * Делегирует запрос к репозиторию настроек и извлекает флаг тёмной темы.
+ *
+ * @param settingsRepository репозиторий для доступа к настройкам темы
+ */
 class GetThemeStateUseCase @Inject constructor(
     private val settingsRepository: SettingsRepository
 ) : GetThemeStateUseCaseContract {
 
+    /**
+     * Выполняет получение текущего состояния темы.
+     *
+     * @return true, если включена тёмная тема; false — если светлая
+     */
     override operator fun invoke(): Boolean {
         return settingsRepository.getThemeSettings().isDarkTheme
     }
